+++ Models/Vector.h
@@ -10,9 +10,9 @@ public:
 	inline Vector();
 	Vector(int size);
 	Vector(int size, const double initial);
-	Vector(double* data, int count);
+	Vector(long double* data, int count);
 	template<int c>
-	Vector(double (&vector)[c]);
+	Vector(long double (&vector)[c]);
 	~Vector();
 	/* Accessors */
 	inline bool isRow();
@@ -21,24 +21,24 @@ public:
 	std::string toString();
 	/* Operations */
 	bool equals(Vector& vector);
-	double dotProduct(Vector& vector);
+	long double dotProduct(Vector& vector);
 	Vector& hadamardProduct(Vector& vector);
 	Vector& transpose();
 	Vector& operator+(Vector& vector);
-	Vector& operator*(double value);
+	Vector& operator*(long double value);
 	Vector& operator-(Vector& vector);
 	Vector& operator*(Vector& vector);	// dot product
 	/* Operator overloads (inline = have to be defined in header file) */
-	inline double& operator()(int index) { return *(data + index); }
-	inline double operator()(int index) const { return *(data + index); };
-	inline double& operator[](int index) { return *(data + index); }
-	inline double operator[](int index) const { return *(data + index); };
+	inline long double& operator()(int index) { return *(data + index); }
+	inline long double operator()(int index) const { return *(data + index); };
+	inline long double& operator[](int index) { return *(data + index); }
+	inline long double operator[](int index) const { return *(data + index); };
 private:
 	bool row = 1;	// by default create row vectors
-	double* data;
+	long double* data;
 
-	void initialize(const double value);
-	void slice(double* elements);
+	void initialize(const long double value);
+	void slice(long double* elements);
 };
 
 #include "Vector.tpp"
